requirements.txt
pandas
tensorflow
sklearn
flask-restful
flask-jsonpify